:recycle: code refactoring
