+++ src/App.tsx
@@ -10,9 +10,9 @@ import { Login } from './containers/Auth/Login/Login';
 import { StaffManagementList } from './containers/StaffManagement/StaffManagementList/StaffManagementList';
 import { StaffManagementTemplate } from './containers/StaffManagement/StaffManagement';
 import { SpeedSendPage } from './components/pages/Template/SpeedSendPage/SpeedSendPage';
-import { SpeedSend } from './containers/Template/SpeedSend/SpeedSend';
+import { SpeedSend } from './containers/Template/Form/SpeedSend/SpeedSend';
 import { HSMPage } from './components/pages/Template/HSMPage/HSMPage';
-import { HSM } from './containers/Template/HSM/HSM';
+import { HSM } from './containers/Template/Form/HSM/HSM';
 import { Chat } from './containers/Chat/Chat';
 import styles from './App.module.css';
 import gqlClient from './config/apolloclient';
