+++ src/components/pages/Template/HSMPage/HSMPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { HSMList } from '../../../../containers/Template/HSM/HSMList/HSMList';
+import { HSMList } from '../../../../containers/Template/List/HSMList/HSMList';
 
 export interface HSMPageProps {}
