+++ src/components/pages/Template/SpeedSendPage/SpeedSendPage.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { SpeedSendPage } from './SpeedSendPage';
-import { SpeedSendList } from '../../../../containers/Template/SpeedSend/SpeedSendList/SpeedSendList';
+import { SpeedSendList } from '../../../../containers/Template/List/SpeedSendList/SpeedSendList';
 
 const wrapper = shallow(<SpeedSendPage />);
