--- src/containers/Template/HSM/HSMList/HSMList.module.css
@@ -1,4 +1,0 @@
-.TemplateIcon {
-  width: 29px;
-  height: 29px;
-}
