--- src/containers/Template/SpeedSend/SpeedSend.module.css
@@ -1,84 +1,0 @@
-.Template {
-  margin: 20px auto;
-  width: 80%;
-  text-align: center;
-  box-shadow: 0 2px 3px #cccccc;
-  border: 1px solid #eeeeee;
-  padding: 10px;
-  box-sizing: border-box;
-}
-
-@media (min-width: 600px) {
-  .Template {
-    width: 500px;
-  }
-}
-
-.DeleteIcon {
-  margin-right: 9px !important;
-}
-
-.DialogText {
-  margin-top: 0px;
-  text-align: center;
-  color: #073f24;
-  font-weight: 400;
-  font-family: 'heebo', sans-serif;
-}
-
-.DeleteButton {
-  margin-left: auto !important;
-}
-
-.Title {
-  margin-left: 24px !important;
-  margin-top: 16px !important;
-  vertical-align: middle;
-  font-weight: 500 !important;
-  color: #073f24;
-}
-
-.Input {
-  display: flex;
-  padding: 8px;
-}
-
-.Label {
-  width: 50%;
-  align-self: center;
-  font-weight: bold;
-}
-
-.TemplateAdd {
-  width: fit-content;
-}
-
-.Form {
-  padding: 16px 16px;
-  width: 470px;
-}
-
-.Buttons {
-  margin-top: 24px;
-  margin-left: 8px;
-  display: flex;
-  justify-content: flex-start;
-}
-
-.Icon {
-  background-color: #eaedec !important;
-  margin-right: 10px !important;
-}
-
-.ButtonsCenter {
-  justify-content: center !important;
-}
-
-.SpeedSendIcon {
-  width: 29px;
-  height: 29px;
-}
-
-.Button {
-  margin-right: 24px !important;
-}
