--- src/containers/Template/SpeedSend/SpeedSend.test.tsx
@@ -1,191 +1,0 @@
-import React from 'react';
-import { render, wait, within, fireEvent, cleanup } from '@testing-library/react';
-import { BrowserRouter as Router } from 'react-router-dom';
-import { MockedProvider } from '@apollo/client/testing';
-import { GET_LANGUAGES } from '../../../graphql/queries/List';
-import {
-  GET_TEMPLATE,
-  GET_TEMPLATES_COUNT,
-  FILTER_TEMPLATES,
-} from '../../../graphql/queries/Template';
-import { CREATE_TEMPLATE } from '../../../graphql/mutations/Template';
-import { SpeedSend } from './SpeedSend';
-import { Switch, Route } from 'react-router-dom';
-import { SpeedSendList } from './SpeedSendList/SpeedSendList';
-
-afterEach(cleanup);
-const mocks = [
-  {
-    request: {
-      query: CREATE_TEMPLATE,
-      variables: {
-        input: {
-          body: 'new Template body',
-          label: 'new Template',
-          languageId: 1,
-          type: 'TEXT',
-        },
-      },
-    },
-    result: {
-      data: {
-        createSessionTemplate: {
-          sessionTemplate: {
-            body: 'new Template body',
-            id: '121',
-            label: 'new Template',
-          },
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: GET_TEMPLATES_COUNT,
-      variables: {
-        filter: {
-          label: '',
-          isHsm: false,
-        },
-      },
-    },
-    result: {
-      data: {
-        countSessionTemplates: 2,
-      },
-    },
-  },
-  {
-    request: {
-      query: FILTER_TEMPLATES,
-      variables: {
-        filter: {
-          label: '',
-          isHsm: false,
-        },
-        opts: {
-          limit: 10,
-          offset: 0,
-          order: 'ASC',
-        },
-      },
-    },
-    result: {
-      data: {
-        sessionTemplates: [
-          {
-            id: '87',
-            label: 'Good message',
-            body: 'Hey There',
-            isReserved: false,
-            isHsm: true,
-          },
-          {
-            id: '94',
-            label: 'Message',
-            body: 'some description',
-            isReserved: false,
-            isHsm: false,
-          },
-        ],
-      },
-    },
-  },
-  {
-    request: {
-      query: GET_LANGUAGES,
-    },
-    result: {
-      data: {
-        languages: [
-          {
-            id: '1',
-            label: 'English (United States)',
-          },
-          {
-            id: '2',
-            label: 'Hindi (India)',
-          },
-        ],
-      },
-    },
-  },
-
-  {
-    request: {
-      query: GET_TEMPLATE,
-      variables: {
-        id: 1,
-      },
-    },
-    result: {
-      data: {
-        sessionTemplate: {
-          sessionTemplate: {
-            id: 1,
-            label: 'important',
-            body: 'important template',
-            isActive: true,
-            language: {
-              id: 1,
-            },
-          },
-        },
-      },
-    },
-  },
-];
-
-test('cancel button should redirect to SpeedSendlist page', async () => {
-  const { container, getByText, unmount } = render(
-    <MockedProvider mocks={mocks} addTypename={false}>
-      <Router>
-        <Switch>
-          <Route path="/speed-send" exact component={SpeedSendList} />
-        </Switch>
-        <SpeedSend match={{ params: { id: 1 } }} />
-      </Router>
-    </MockedProvider>
-  );
-  await wait();
-  const { queryByText } = within(container.querySelector('form'));
-  const button = queryByText('Cancel');
-  fireEvent.click(button);
-  expect(getByText('Loading...')).toBeInTheDocument();
-  await wait();
-  expect(getByText('Speed sends')).toBeInTheDocument();
-  unmount();
-});
-
-describe('Save Button', () => {
-  test('save button should add a new template', async () => {
-    const { container } = render(
-      <MockedProvider mocks={mocks} addTypename={false}>
-        <Router>
-          <SpeedSend match={{ params: { id: null } }} />
-          <Switch>
-            <Route path="/speed-send" exact component={SpeedSendList} />
-          </Switch>
-        </Router>
-      </MockedProvider>
-    );
-
-    await wait();
-    fireEvent.change(container.querySelector('input[name="label"]'), {
-      target: { value: 'new Template' },
-    });
-    fireEvent.change(container.querySelector('textarea[name="body"]'), {
-      target: { value: 'new Template body' },
-    });
-
-    fireEvent.change(container.querySelector('input[name="languageId"]'), {
-      target: { value: 1 },
-    });
-    const { queryByText } = within(container.querySelector('form'));
-    const button = queryByText('Save');
-    fireEvent.click(button);
-    await wait();
-    const { getByText } = within(container.querySelector('tbody'));
-    expect(getByText('Good message')).toBeInTheDocument();
-  });
-});
