--- src/containers/Template/SpeedSend/SpeedSendList/SpeedSendList.module.css
@@ -1,4 +1,0 @@
-.SpeedSendIcon {
-  width: 29px;
-  height: 29px;
-}
