--- src/containers/Template/Template.module.css
@@ -1,34 +1,0 @@
-.Label {
-  width: 255px;
-}
-
-.Body {
-  width: 650px;
-}
-
-.Actions {
-  width: 100px;
-  justify-content: flex-end;
-  margin-left: auto;
-}
-
-.LabelText {
-  font-family: 'heebo', sans-serif;
-  display: flex;
-  font-weight: 500;
-  font-size: 20px;
-  color: #073f24;
-  margin-top: 12px;
-  margin-bottom: 12px;
-  line-height: 1;
-}
-
-.TableText {
-  margin-right: 23px;
-  text-align: left;
-  padding: 12px 0px;
-  font-size: 16px;
-  color: #93a29b;
-  line-height: 1.25;
-  margin: 0;
-}
